Refactor: 제품 상세 JS 파일의 댓글 기능 모듈화(comment.js로 분리)
