+++ src/main/resources/static/assets/js/product-detail.js
@@ -1,9 +1,8 @@
+import * as comment from './comment.js';
+
 /******************** 전역 변수 **********************/
 let productId = 0;
 let globalSort = 'createDt,DESC';
-let initComment = '';
-
-const list = document.querySelector('.reviewImage-modal');
 
 /******************** 함수 **********************/
 // 리뷰 리스트 조회 데이터 받아서 정렬
@@ -181,7 +180,7 @@ const fnGetReviewDetail = (evt) => {
     .then((response) => {
 
         fnShowReviewDetail(response.data.review);
-        fnShowCommentList(response.data.commentList);
+        comment.fnShowCommentList(response.data.commentList);
 
         $('#review-detail').modal('show');
     })
@@ -241,112 +240,6 @@ const fnShowReviewDetail = (review) => {
 
 }
 
-// 리뷰 댓글 데이터 화면에 표시
-const fnShowCommentList = (commentList) => {
-
-    let loginId = $('.login-id').val();
-    let commentContainer = $('.content-3');
-    let writerName = $('.writerImage-modal ~ p').text();
-
-    let str = ''
-    commentList.forEach((comment) => {
-
-        if(comment.depth === 0) {
-            str += '<div class="comment-modal" data-group-id="' + comment.groupId + '" data-comment-id="' + comment.commentId + '">';
-        } else {
-            str += '<div class="comment-modal reply" data-group-id="' + comment.groupId + '" data-comment-id="' + comment.commentId + '">';
-        }
-        str += '<div class="comment-writer-section">';
-        if(comment.member.profileImage !== null) {
-            str += '<div class="comment-writer"><img src="' + comment.member.profileImage + '"></div>';
-        } else {
-            str += '<div class="comment-writer"><img src="/assets/img/default-profile.jpg"></div>';
-        }
-        str += '</div>';
-        str += '<div class="w-100">';
-        str += '<div class="comment-profile">';
-        if(comment.member.nickname === writerName) {
-            str += '<p style="color: #D9232D;">글쓴이</p>';
-        } else {
-            str += '<p>' + comment.member.nickname + '</p>';
-        }
-        str += '<p>' + fnFormatDate(comment.createDt) + '</p>';
-        str += '</div>';
-
-        if(loginId !== undefined) {
-
-            if(comment.member.memberId === Number(loginId)) {
-
-                if(comment.depth === 0) {
-                    str += '<div class="comment-content"><p>' + comment.content + '</p>';
-                    str += '<div class="d-flex gap-1">';
-                    str += '<p class="btn-reply" style="align-content: end;">댓글달기</p>';
-                    str += '<div>';
-                    str += '<i class="fa-solid fa-pen-to-square icon-comment-update" style="align-content: center;"></i> | ';
-                    str += '<i class="fa-solid fa-trash icon-comment-delete" style="align-content: center;"></i>';
-                    str += '</div>';
-                    str += '</div>';
-                    str += '</div>';
-                } else {
-                    str += '<div class="comment-content">';
-                    str += '<p>' + comment.content + '</p>';
-                    str += '<div>';
-                    str += '<i class="fa-solid fa-pen-to-square icon-comment-update" style="align-content: center;"></i>';
-                    str += ' | ';
-                    str += '<i class="fa-solid fa-trash icon-comment-delete" style="align-content: center;"></i>';
-                    str += '</div>';
-                    str += '</div>';
-                }
-            } else {
-
-                if(comment.depth === 0) {
-                    str += '<div class="comment-content"><p>' + comment.content + '</p><p class="btn-reply">댓글달기</p></div>';
-                } else {
-                    str += '<div class="comment-content"><p>' + comment.content + '</p></div>';
-                }
-            }
-        } else {
-            str += '<div class="comment-content"><p>' + comment.content + '</p></div>';
-        }
-        str += '</div>';
-        str += '</div>';
-
-    });
-
-    commentContainer.empty();
-    commentContainer.html(str);
-
-}
-
-// 날짜 포맷
-const fnFormatDate = (datetime) => {
-    const date = new Date(datetime);
-    const now = new Date();
-    const diffMs = now - date; // 시간 차이 (밀리초 단위)
-    const diffMin = Math.floor(diffMs / 60000);
-
-    if(diffMin < 1) {
-        return '방금 전';
-    }
-
-    if(diffMin < 60) {
-        return diffMin + '분 전';
-    }
-
-    if(diffMin < 120) {
-        return '1시간 전';
-    }
-
-    // 1시간 이상 경과되었을 경우..
-    const yyyy = date.getFullYear();
-    const mm = String(date.getMonth()).padStart(2, '0');
-    const dd = String(date.getDate()).padStart(2, '0');
-    const hh = String(date.getHours()).padStart(2, '0');
-    const mi = String(date.getMinutes()).padStart(2, '0');
-
-    return yyyy + '-' + mm + '-' + dd + ' ' + hh + ':' + mi;
-}
-
 // csrf 토큰 가져오기
 const fnGetCsrfToken = () => {
     return $('meta[name="csrf-token"]').attr('content');
@@ -376,245 +269,9 @@ const fnDeleteReview = (evt) => {
     });
 }
 
-// 대댓글 선택
-const fnSelectReply = (item, isSelected) => {
-
-    let commentInput = $('.comment-input input');
-
-    fnChangeCommentBtn('insert');
-
-    if(isSelected) {
-
-        let comment = item.closest('.comment-modal');
-
-        $('.comment-modal').not(comment).removeClass('selected-comment');
-        comment.addClass('selected-comment');
-        commentInput.val('');
-        commentInput.attr('placeholder', '대댓글을 입력해주세요');
-
-    } else {
-
-        let comment = item;
-
-        comment.removeClass('selected-comment');
-        commentInput.val('');
-        commentInput.attr('placeholder', '댓글을 입력해주세요');
-    }
-}
-
-// 작성 댓글 표시
-const fnShowComment = (comment) => {
-
-    let commentArea = $('.content-3');
-    let isReply = comment.depth === 0;
-    let writerName = $('.writerImage-modal ~ p').text();
-
-    let commentContainer
-        = isReply ? commentArea : $('[data-group-id="' + comment.groupId + '"]').last();
-
-    let loginId = $('.login-id').val();
-
-    let str = '';
-    if(isReply) {
-        str += '<div class="comment-modal" data-group-id="' + comment.groupId + '" data-comment-id="' + comment.commentId + '">';
-    } else {
-        str += '<div class="comment-modal reply" data-group-id="' + comment.groupId + '" data-comment-id="' + comment.commentId + '">';
-    }
-
-    str += '<div class="comment-writer-section">';
-    if(comment.member.profileImage !== null) {
-        str += '<div class="comment-writer"><img src="' + comment.member.profileImage + '"></div>';
-    } else {
-        str += '<div class="comment-writer"><img src="/assets/img/default-profile.jpg"></div>';
-    }
-    str += '</div>';
-    str += '<div class="w-100">';
-    str += '<div class="comment-profile">';
-    if(comment.member.nickname === writerName) {
-        str += '<p style="color: #D9232D;">글쓴이</p>';
-    } else {
-        str += '<p>' + comment.member.nickname + '</p>';
-    }
-    str += '<p>' + fnFormatDate(comment.createDt) + '</p>';
-    str += '</div>';
-
-    if(loginId !== undefined) {
-
-        if(comment.member.memberId === Number(loginId)) {
-
-            if(comment.depth === 0) {
-                str += '<div class="comment-content"><p>' + comment.content + '</p>';
-                str += '<div class="d-flex gap-1">';
-                str += '<p class="btn-reply" style="align-content: end;">댓글달기</p>';
-                str += '<div>';
-                str += '<i class="fa-solid fa-pen-to-square icon-comment-update" style="align-content: center;"></i> | ';
-                str += '<i class="fa-solid fa-trash icon-comment-delete" style="align-content: center;"></i>';
-                str += '</div>';
-                str += '</div>';
-                str += '</div>';
-            } else {
-                str += '<div class="comment-content">';
-                str += '<p>' + comment.content + '</p>';
-                str += '<div>';
-                str += '<i class="fa-solid fa-pen-to-square icon-comment-update" style="align-content: center;"></i>';
-                str += ' | ';
-                str += '<i class="fa-solid fa-trash icon-comment-delete" style="align-content: center;"></i>';
-                str += '</div>';
-                str += '</div>';
-            }
-        } else {
-
-            if(comment.depth === 0) {
-                str += '<div class="comment-content"><p>' + comment.content + '</p><p class="btn-reply">댓글달기</p></div>';
-            } else {
-                str += '<div class="comment-content"><p>' + comment.content + '</p></div>';
-            }
-        }
-    } else {
-        str += '<div class="comment-content"><p>' + comment.content + '</p></div>';
-    }
-    str += '</div>';
-    str += '</div>';
-
-    if(isReply) { // 댓글인 경우 맨 마지막에 댓글 추가 및 스크롤 맨 아래로 내리기
-        commentContainer.append(str);
-        commentArea.scrollTop(commentArea[0].scrollHeight);
-    } else { // 대댓글인 경우 같은 group_id를 가진 요소 다음에 댓글 추가 및 스크롤 해당 댓글의 원글로 이동
-        commentContainer.after(str);
-    }
-
-    $('.selected-comment').removeClass('selected-comment');
-    $('.comment-input input').val('');
-
-}
-
-// 댓글 작성
-const fnWriteComment = (selectedComment) => {
-
-    const isReply = selectedComment.length !== 0;
-
-    let content = $('.comment-input input').val().trim();
-    let reviewId = $('.review-id').val();
-
-    if(content === '') {
-        alert('댓글을 입력해주세요.');
-        return;
-    }
-
-    const commentDTO = {
-        reviewId,
-        content,
-        depth: isReply ? 1 : 0,
-        groupId: isReply ? selectedComment.data('commentId') : undefined,
-        state: false
-    }
-
-    axios.post('/comment/insertComment', commentDTO,
-        {
-            headers: {
-                'Content-Type': 'application/json',
-                'X-XSRF-TOKEN': fnGetCsrfToken()
-            }
-        })
-    .then((response) => {
-        if(response.data.success) {
-            fnShowComment(response.data.comment);
-        }
-    })
-    .catch((error) => {
-        alert(response.data.message);
-    });
-
-}
-
-// 댓글 수정 입력 표시
-const fnSetUpdateComment = (selectedComment) => {
-
-    // 수정할 댓글 내용을 가져와서 input에 설정
-    let commentId = selectedComment.closest('.comment-modal').data('commentId');
-    let content = selectedComment.closest('.comment-content').find('p').first().text();
-    let commentInput = $('.comment-input input');
-
-    fnChangeCommentBtn('update');
-    commentInput.val(content);
-    commentInput.attr('placeholder', '수정할 댓글을 입력해주세요.');
-
-    initComment = content;
-
-    $('.update-comment-id').val(commentId);
-}
-
-// 댓글 버튼 타입 변경
-const fnChangeCommentBtn = (type) => {
-    if(type === 'insert') {
-        $('.btn-comment').css('display', '');
-        $('.btn-comment-update').css('display', 'none');
-    } else {
-        $('.btn-comment').css('display', 'none');
-        $('.btn-comment-update').css('display', '');
-    }
-}
-
-// 수정한 댓글 표시
-const fnShowUpdateComment = (comment) => {
-
-    let commentId = comment.commentId;
-    let commentContent = $('[data-comment-id="' + commentId + '"]').find('.comment-content p').first();
-    let commentInput = $('.comment-input input');
-
-    commentContent.text(comment.content);
-    commentInput.val('');
-    commentInput.attr('placeholder', '댓글을 입력해주세요.');
-
-    fnChangeCommentBtn('insert');
-}
-
-// 댓글 수정
-const fnUpdateComment = () => {
-
-    let content = $('.comment-input input');
-    let commentId = $('.update-comment-id').val();
-
-    // 빈값 검사
-    if(content.val().trim() === '') {
-        alert('댓글을 입력해주세요.');
-        return;
-    }
-
-    // 초기 댓글 내용과 비교
-    if(content.val().trim() === initComment) {
-        alert('댓글이 수정되었습니다.');
-        content.val('');
-        content.attr('placeholder', '댓글을 작성해주세요.');
-        fnChangeCommentBtn('insert');
-        return;
-    }
-
-    const commentDTO = {
-        commentId: commentId,
-        content: content.val().trim()
-    }
-
-    axios.put('/comment/updateComment', commentDTO,
-        {
-            headers: {
-                'Content-Type': 'application/json',
-                'X-XSRF-TOKEN': fnGetCsrfToken()
-            }
-        })
-        .then((response) => {
-            if(response.data.success) {
-                alert('댓글이 수정되었습니다.');
-                fnShowUpdateComment(response.data.comment);
-            }
-        })
-        .catch((error) => {
-            alert(response.data.message);
-        });
-}
-
-
 /******************** 이벤트 **********************/
+window.fnGetReviewList = fnGetReviewList;
+
 $(document).ready(() => {
     const urlParams = new URL(location.href).searchParams;
     productId = urlParams.get('productId');
@@ -654,30 +311,3 @@ $(document).on('click', '.btn-update', (evt) => {
     location.href = "/review/reviewUpdate.page?reviewId=" + reviewId + '&productId=' + productId;
 });
 
-// 대댓글 달 원글 선택
-$(document).on('click', '.btn-reply', (evt) => {
-    evt.stopPropagation();
-    fnSelectReply($(evt.currentTarget), true);
-});
-
-// 대댓글 달 원글 선택 취소
-$(document).on('click', '.comment-modal', (evt) => {
-    evt.stopPropagation();
-    fnSelectReply($(evt.currentTarget), false);
-});
-
-// 댓글 작성
-$('.btn-comment').on('click', () => {
-    fnWriteComment($('.selected-comment'));
-});
-
-// 댓글 수정 input 표시
-$(document).on('click', '.icon-comment-update', (evt) => {
-    evt.stopPropagation();
-    fnSetUpdateComment($(evt.currentTarget));
-});
-
-// 댓글 수정
-$('.btn-comment-update').on('click', (evt) => {
-    fnUpdateComment();
-});
