+++ src/main/resources/templates/product/product-details.html
@@ -133,7 +133,6 @@
 <!-- 공통 script 분리 -->
 <div th:replace="~{scripts/scripts::scripts}"></div>
 
-<script src="/assets/js/product-detail.js"></script>
-
+<script type="module" src="/assets/js/product-detail.js"></script>
 </body>
 </html>
